# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "example": "admin access required",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "example": "subscription created successfully",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "type": "object"
      },
      "PushSubscriptionResponse": {
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "endpoint": {
            "type": "string"
          },
          "lastUsed": {
            "type": "string"
          },
          "subscriptionId": {
            "type": "integer"
          },
          "userAgent": {
            "type": "string"
          }
        },
        "required": [
          "createdAt",
          "endpoint",
          "lastUsed",
          "subscriptionId",
          "userAgent"
        ],
        "type": "object"
      },
...
      "SubscribePayload": {
        "properties": {
          "endpoint": {
            "type": "string"
          },
          "keys": {
            "properties": {
              "auth": {
                "type": "string"
              },
              "p256dh": {
                "type": "string"
              }
            },
            "required": [
              "auth",
              "p256dh"
            ],
            "type": "object"
          }
        },
        "required": [
          "endpoint",
          "keys"
        ],
        "type": "object"
      },
...
      "TestNotificationResponse": {
        "properties": {
          "message": {
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/TestNotificationResult"
            },
            "type": "array"
          },
          "subscriptionCount": {
            "type": "integer"
          }
        },
        "required": [
          "message",
          "results",
          "subscriptionCount"
        ],
        "type": "object"
      },
      "TestNotificationResult": {
        "properties": {
          "endpoint": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "httpStatus": {
            "type": "integer"
          },
          "subscriptionId": {
            "type": "integer"
          },
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "endpoint",
          "error",
          "httpStatus",
          "subscriptionId",
          "success"
        ],
        "type": "object"
      },
      "UnsubscribePayload": {
        "properties": {
          "endpoint": {
            "type": "string"
          }
        },
        "required": [
          "endpoint"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "Raw JWT token from POST /users/login. Sent verbatim with no \"Bearer \" prefix.",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "REST API for managing animals, enclosures and care tasks.\n\nOnly v2 routes are documented here. v1 remains served at /api/v1 for\nbackwards compatibility but is not part of this contract.",
    "title": "Animal Family API",
    "version": "2.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/notifications/subscribe": {
      "post": {
        "operationId": "subscribeToPush",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscribePayload"
              }
            }
          },
          "description": "Browser push subscription",
          "required": true,
          "x-originalParamName": "subscription"
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Register a push subscription",
        "tags": [
          "notifications"
        ]
      }
    },
    "/notifications/subscriptions": {
      "get": {
        "operationId": "listPushSubscriptions",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PushSubscriptionResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List the caller's push subscriptions",
        "tags": [
          "notifications"
        ]
      }
    },
    "/notifications/test": {
      "post": {
        "description": "Delivers a test message to every subscription the caller has registered, reporting the outcome for each.",
        "operationId": "sendTestNotification",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestNotificationResponse"
                }
              }
            },
            "description": "OK"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Send a test push notification",
        "tags": [
          "notifications"
        ]
      }
    },
    "/notifications/unsubscribe": {
      "post": {
        "operationId": "unsubscribeFromPush",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnsubscribePayload"
              }
            }
          },
          "description": "Endpoint to remove",
          "required": true,
          "x-originalParamName": "subscription"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Remove a push subscription",
        "tags": [
          "notifications"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Relative to the API host, configured per environment.",
      "url": "/api/v2"
    }
  ]
}